+++ BlueDBOnDisk/src/test/java/io/bluedb/TestUtils.java
@@ -1,0 +1,30 @@
+package io.bluedb;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Comparator;
+
+public class TestUtils {
+	public static void recursiveDelete(File file) {
+		try {
+			Files.walk(file.toPath())
+			.sorted(Comparator.reverseOrder())
+			.map(Path::toFile)
+			.forEach(File::delete);
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		if (!file.exists()) {
+			return;
+		} else if (file.isDirectory()) {
+			for (File f: file.listFiles()) {
+				recursiveDelete(f);
+			}
+			file.delete();
+		} else {
+			file.delete();
+		}
+	}
+}
