+++ BlueDBOnDisk/src/test/java/io/bluedb/zip/ZipUtilsTest.java
@@ -1,13 +1,26 @@
 package io.bluedb.zip;
 
-import static org.junit.Assert.*;
-import java.io.IOException;
+import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import org.junit.Test;
+import junit.framework.TestCase;
 
-public class ZipUtilsTest {
+public class ZipUtilsTest extends TestCase {
+
+	Path tempFolder;
+
+	@Override
+	public void setUp() throws Exception {
+		tempFolder = Files.createTempDirectory(this.getClass().getSimpleName());
+		tempFolder.toFile().deleteOnExit();
+	}
+
+	@Override
+	public void tearDown() {
+		recursiveDelete(tempFolder.toFile());
+	}
 
 	@Test
 	public void test_constructor() {
@@ -15,44 +28,49 @@ public class ZipUtilsTest {
 	}
 
 	@Test
-	public void test() {
-		try {
-			Path tempFolder = Files.createTempDirectory(this.getClass().getSimpleName());
-			tempFolder.toFile().deleteOnExit();
-			
-			Path folderToCopy = Paths.get(tempFolder.toString(), "folder");
-			Path fileInFolderToCopy = Paths.get(folderToCopy.toString(), "file");
-			Path subfolderToCopy = Paths.get(folderToCopy.toString(), "subfolder");
-			Path fileInSubfolderToCopy = Paths.get(subfolderToCopy.toString(), "subfolder_file");
-			
-			folderToCopy.toFile().mkdirs();
-			fileInFolderToCopy.toFile().createNewFile();
-			subfolderToCopy.toFile().mkdirs();
-			fileInSubfolderToCopy.toFile().createNewFile();
-			
-			Path zippedPath = Paths.get(tempFolder.toString(), "test.zip");
-			ZipUtils.zipFile(folderToCopy, zippedPath);
-	
-			Path unzippedFolder = Paths.get(tempFolder.toString(), "unzipped");
-			ZipUtils.extractFiles(zippedPath, unzippedFolder);
-			
-			Path unzippedFolderToCopy = translatePath(folderToCopy, tempFolder, unzippedFolder);
-			Path unzippedFileInFolderToCopy = translatePath(fileInFolderToCopy, tempFolder, unzippedFolder);
-			Path unzippedSubfolderToCopy = translatePath(subfolderToCopy, tempFolder, unzippedFolder);
-			Path unzippedFileInSubfolderToCopy = translatePath(fileInSubfolderToCopy, tempFolder, unzippedFolder);
-			
-			assertTrue(unzippedFolderToCopy.toFile().exists());
-			assertTrue(unzippedFileInFolderToCopy.toFile().exists());
-			assertTrue(unzippedSubfolderToCopy.toFile().exists());
-			assertTrue(unzippedFileInSubfolderToCopy.toFile().exists());
-		} catch (IOException e) {
-			e.printStackTrace();
-			fail();
-		}
+	public void test() throws Exception {
+		Path folderToCopy = Paths.get(tempFolder.toString(), "folder");
+		Path fileInFolderToCopy = Paths.get(folderToCopy.toString(), "file");
+		Path subfolderToCopy = Paths.get(folderToCopy.toString(), "subfolder");
+		Path fileInSubfolderToCopy = Paths.get(subfolderToCopy.toString(), "subfolder_file");
+		
+		folderToCopy.toFile().mkdirs();
+		fileInFolderToCopy.toFile().createNewFile();
+		subfolderToCopy.toFile().mkdirs();
+		fileInSubfolderToCopy.toFile().createNewFile();
+		
+		Path zippedPath = Paths.get(tempFolder.toString(), "test.zip");
+		ZipUtils.zipFile(folderToCopy, zippedPath);
+
+		Path unzippedFolder = Paths.get(tempFolder.toString(), "unzipped");
+		ZipUtils.extractFiles(zippedPath, unzippedFolder);
+		
+		Path unzippedFolderToCopy = translatePath(folderToCopy, tempFolder, unzippedFolder);
+		Path unzippedFileInFolderToCopy = translatePath(fileInFolderToCopy, tempFolder, unzippedFolder);
+		Path unzippedSubfolderToCopy = translatePath(subfolderToCopy, tempFolder, unzippedFolder);
+		Path unzippedFileInSubfolderToCopy = translatePath(fileInSubfolderToCopy, tempFolder, unzippedFolder);
+		
+		assertTrue(unzippedFolderToCopy.toFile().exists());
+		assertTrue(unzippedFileInFolderToCopy.toFile().exists());
+		assertTrue(unzippedSubfolderToCopy.toFile().exists());
+		assertTrue(unzippedFileInSubfolderToCopy.toFile().exists());
 	}
 
 	public Path translatePath(Path targetPath, Path from, Path to) {
 		Path relativePath = from.relativize(targetPath);
 		return Paths.get(to.toString(), relativePath.toString());
+	}
+
+	private void recursiveDelete(File file) {
+		if (!file.exists()) {
+			return;
+		} else if (file.isDirectory()) {
+			for (File f: file.listFiles()) {
+				recursiveDelete(f);
+			}
+			file.delete();
+		} else {
+			file.delete();
+		}
 	}
 }
